+++ backend/src/test/java/com/tissue/helper/ControllerTestHelper.java
@@ -31,8 +31,8 @@ import com.tissue.api.position.service.query.PositionQueryService;
 import com.tissue.api.review.domain.repository.IssueReviewerRepository;
 import com.tissue.api.review.presentation.controller.ReviewController;
 import com.tissue.api.review.presentation.controller.ReviewerController;
-import com.tissue.api.review.service.ReviewCommandService;
-import com.tissue.api.review.service.ReviewerCommandService;
+import com.tissue.api.review.service.command.ReviewCommandService;
+import com.tissue.api.review.service.command.ReviewerCommandService;
 import com.tissue.api.security.authentication.presentation.controller.AuthenticationController;
 import com.tissue.api.security.authentication.service.AuthenticationService;
 import com.tissue.api.security.session.SessionManager;
