+++ backend/src/test/java/com/tissue/helper/ServiceIntegrationTestHelper.java
@@ -20,8 +20,8 @@ import com.tissue.api.position.service.command.PositionCommandService;
 import com.tissue.api.position.service.query.PositionQueryService;
 import com.tissue.api.review.domain.repository.IssueReviewerRepository;
 import com.tissue.api.review.domain.repository.ReviewRepository;
-import com.tissue.api.review.service.ReviewCommandService;
-import com.tissue.api.review.service.ReviewerCommandService;
+import com.tissue.api.review.service.command.ReviewCommandService;
+import com.tissue.api.review.service.command.ReviewerCommandService;
 import com.tissue.api.security.PasswordEncoder;
 import com.tissue.api.team.domain.respository.TeamRepository;
 import com.tissue.api.team.service.command.TeamCommandService;
